requirements.txt
altgraph==0.17.2
future==0.18.2
importlib-metadata==4.8.1
pefile==2021.9.3
psutil==5.8.0
py-cpuinfo==8.0.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pywin32==301
pywin32-ctypes==0.2.0
typing-extensions==3.10.0.2
zipp==3.6.0
